moved celery config into extensions file
