+++ app/util/extensions.py
@@ -1,3 +1,21 @@
 from flask_sqlalchemy import SQLAlchemy
+from celery import Celery
 
 db = SQLAlchemy()
+
+def create_celery_app(app=None):
+    app = app or create_app()
+    celery = Celery(__name__, broker=app.config['CELERY_BROKER_URL'])
+    celery.conf.update(app.config)
+    TaskBase = celery.Task
+
+    class ContextTask(TaskBase):
+        abstract = True
+
+        def __call__(self, *args, **kwargs):
+            with app.app_context():
+                return TaskBase.__call__(self, *args, **kwargs)
+
+    celery.Task = ContextTask
+    celery.app = app
+    return celery
